requirements.txt
django==1.8.5
djangorestframework==3.3.3
django-cors-headers==1.1.0
django-filter==0.13.0
